Use pg-config to try and determine the location of libpq
On SOME PLATFORMS (not going to name any names here, but it rhymes with
derpdows), PostgreSQL will put libpq in a non-default-search-path
location. If that platform also doesn't provide a standard env var to
configure where to look, there's no way to actually find it.

If you have PG on your machine, we can assume you have `pg-config` on your
path. This will tell us where the heck to look.