Try to avoid adding `/usr/local/lib` to the search path on Mac
So this is a bit weird to do, and it's somewhat of a hack. However since
https://github.com/rust-lang/cargo/pull/2763, cargo will now put any `-L`
arguments onto `DYLD_LIBRARY_PATH` when using `cargo test` or `cargo run`. For
installations using Homebrew, this means that we will try to stick
`/usr/local/lib` in `DYLD_LIBRARY_PATH`. If there's anything in there that
conflicts with system libraries (such as `libJPEG`) everything blows up.

It is certainly possible for this configuration issue to affect platforms other
than Mac, but this is not meant to be a catch-all solution for platform
configuration problems. For Mac users though, this is a very common setup that
is easy to resolve. After this commit, the directory will be something like
`/usr/local/Cellar/postgresql/9.5.3/lib` and not `/usr/local/lib`.