Refine check_compiles function
It now uses the compiler configuration that is used with the normal
program _plus_ appropriate flags to make sure implicit function
declarations cause compilation errors (instead of silently working).
Failing the compilation in this case is intended, as this is used to
detect whether or not a feature is supported on a platform.